+++ dongmanee/src/main/java/com/dongmanee/domain/email/service/EmailServiceImpl.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Service;
 import com.dongmanee.domain.email.exception.EmailSendingException;
 import com.dongmanee.domain.email.exception.EmailVerifiedException;
 import com.dongmanee.domain.email.utils.EmailRedisUtils;
-import com.dongmanee.domain.member.dao.MemberRepository;
+import com.dongmanee.domain.member.dao.jpa.MemberJpaRepository;
 import com.dongmanee.domain.member.exception.DuplicateEmailException;
 import com.dongmanee.global.utils.AuthCodeProvider;
 
@@ -27,7 +27,7 @@ import lombok.extern.slf4j.Slf4j;
 public class EmailServiceImpl implements EmailService {
 	private final JavaMailSender emailSender;
 	private final EmailRedisUtils emailRedis;
-	private final MemberRepository memberRepository;
+	private final MemberJpaRepository memberJpaRepository;
 	private final AuthCodeProvider authCodeProvider;
 
 	@Value("${spring.mail.username}")
@@ -46,7 +46,7 @@ public class EmailServiceImpl implements EmailService {
 	 */
 	@Override
 	public void sendSingUpEmailAuthCode(String toEmail) {
-		if (memberRepository.existsByEmail(toEmail)) {
+		if (memberJpaRepository.existsByEmail(toEmail)) {
 			throw new DuplicateEmailException("이미 가입한 이메일입니다.");
 		}
