+++ dongmanee/src/main/java/com/dongmanee/domain/member/dao/MemberRepository.java
@@ -2,12 +2,10 @@ package com.dongmanee.domain.member.dao;
 
 import java.util.Optional;
 
-import org.springframework.data.jpa.repository.JpaRepository;
-
+import com.dongmanee.domain.member.dao.jpa.MemberJpaRepository;
 import com.dongmanee.domain.member.domain.Member;
-import com.dongmanee.domain.member.service.port.SignUpServiceMemberRepository;
 
-public interface MemberRepository extends JpaRepository<Member, Long>, SignUpServiceMemberRepository {
+public interface MemberRepository extends MemberJpaRepository {
 	boolean existsByStudentId(String studentId);
 
 	boolean existsByEmail(String email);
