+++ dongmanee/src/main/java/com/dongmanee/domain/member/service/MemberServiceImpl.java
@@ -2,7 +2,7 @@ package com.dongmanee.domain.member.service;
 
 import org.springframework.stereotype.Service;
 
-import com.dongmanee.domain.member.dao.MemberRepository;
+import com.dongmanee.domain.member.dao.jpa.MemberJpaRepository;
 import com.dongmanee.domain.member.domain.Member;
 import com.dongmanee.domain.member.exception.MemberNotFoundException;
 
@@ -11,10 +11,10 @@ import lombok.RequiredArgsConstructor;
 @Service
 @RequiredArgsConstructor
 public class MemberServiceImpl implements MemberService {
-	private final MemberRepository memberRepository;
+	private final MemberJpaRepository memberJpaRepository;
 
 	@Override
 	public Member findById(Long id) {
-		return memberRepository.findById(id).orElseThrow(MemberNotFoundException::new);
+		return memberJpaRepository.findById(id).orElseThrow(MemberNotFoundException::new);
 	}
 }
