--- dongmanee/src/main/java/com/dongmanee/domain/member/service/port/SignUpServiceAuthProviderRepository.java
@@ -1,9 +1,0 @@
-package com.dongmanee.domain.member.service.port;
-
-import java.util.Optional;
-
-import com.dongmanee.domain.security.domain.AuthProvider;
-
-public interface SignUpServiceAuthProviderRepository {
-	Optional<AuthProvider> findByAuthProviderAndExternalProviderId(String provider, Long externalProviderId);
-}
